fix: active state event listeners
