+++ package.json
@@ -124,6 +124,7 @@
     "@rm/types": "*",
     "@sentry/vite-plugin": "2.2.1",
     "@types/dlv": "^1.1.2",
+    "@types/node": "^18",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.0.9",
     "@vitejs/plugin-react": "4.0.0",
